requirements.txt
gradio==5.47.0
openai==1.109.1
python-dotenv==1.1.1
pypdf==6.1.0
